requirements.txt
streamlit
numpy
matplotlib
sympy
fpdf
Pillow